Move SHOW CREATE formatters to a separate library (#52592)
