+++ ydb/core/sys_view/show_create/create_external_table_formatter.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "formatters_common.h"
+#include "formatters/formatters_common.h"
 
 #include <ydb/core/protos/flat_scheme_op.pb.h>
